refactor: change directory for storybook
