+++ client/src/component/atom/Button/Button.stories.tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+
+import Button from '.'
+
+export default {
+  title: 'Atoms/Button',
+  component: Button,
+}
+
+export const Default = () => {
+  return <Button>hi</Button>
+}
