+++ client/src/component/atom/Button/index.ts
@@ -1,0 +1,23 @@
+import { ReactChild } from 'react'
+
+export { default } from './Button'
+
+export namespace ButtonType {
+  export interface StyleAttributes extends Object {
+    height?: string
+    width?: string
+    margin?: string
+    padding?: string
+    rounded?: boolean
+    border?: string
+    backgroundColor?: string
+    disabled?: boolean
+    hover?: boolean
+  }
+
+  export interface Props {
+    customStyle?: StyleAttributes
+    children?: ReactChild
+    onClick?: () => void
+  }
+}
